Issue 3366 - HAGroup: Agent to use hagroup from the node instead of HAAttribute
