+++ api/path_node_test.go
@@ -64,7 +64,7 @@ func Test_FindHDForOutput1(t *testing.T) {
 
 	theOrg := "myorg"
 
-	_, err = persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "device", false, theOrg, "apattern", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	_, err = persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "device", theOrg, "apattern", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("failed to create persisted device, error %v", err)
 	}
@@ -95,7 +95,7 @@ func Test_FindHDForOutput2(t *testing.T) {
 
 	theOrg := "myorg"
 
-	_, err = persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "", false, theOrg, "otherorg/apattern", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	_, err = persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "", theOrg, "otherorg/apattern", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("failed to create persisted device, error %v", err)
 	}
@@ -713,7 +713,7 @@ func Test_DeleteHorizonDevice_success(t *testing.T) {
 	myPattern := "testPattern"
 	device := getBasicDevice(myOrg, myPattern)
 
-	_, err = persistence.SaveNewExchangeDevice(db, *device.Id, *device.Token, *device.Name, "cluster", false, *device.Org, *device.Pattern, persistence.CONFIGSTATE_CONFIGURED, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	_, err = persistence.SaveNewExchangeDevice(db, *device.Id, *device.Token, *device.Name, "cluster", *device.Org, *device.Pattern, persistence.CONFIGSTATE_CONFIGURED, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("unexpected error creating device %v", err)
 	}
@@ -752,7 +752,7 @@ func Test_DeleteHorizonDevice_fail1(t *testing.T) {
 	myPattern := "testPattern"
 	device := getBasicDevice(myOrg, myPattern)
 
-	_, err = persistence.SaveNewExchangeDevice(db, *device.Id, *device.Token, *device.Name, "device", false, *device.Org, *device.Pattern, persistence.CONFIGSTATE_UNCONFIGURED, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	_, err = persistence.SaveNewExchangeDevice(db, *device.Id, *device.Token, *device.Name, "device", *device.Org, *device.Pattern, persistence.CONFIGSTATE_UNCONFIGURED, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("unexpected error creating device %v", err)
 	}
@@ -793,7 +793,7 @@ func Test_PatchHorizonDevice_fail1(t *testing.T) {
 	myPattern := "testPattern"
 	device := getBasicDevice(myOrg, myPattern)
 
-	_, err = persistence.SaveNewExchangeDevice(db, *device.Id, *device.Token, *device.Name, "", false, *device.Org, *device.Pattern, persistence.CONFIGSTATE_CONFIGURED, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	_, err = persistence.SaveNewExchangeDevice(db, *device.Id, *device.Token, *device.Name, "", *device.Org, *device.Pattern, persistence.CONFIGSTATE_CONFIGURED, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("unexpected error creating device %v", err)
 	}
@@ -827,7 +827,7 @@ func getBasicDevice(org string, pattern string) *HorizonDevice {
 	myId := "testid"
 	myName := "testName"
 	myToken := "testToken"
-	myHA := false
+	myHAGroup := "ha_group_name"
 
 	hd := &HorizonDevice{
 		Id:      &myId,
@@ -835,7 +835,7 @@ func getBasicDevice(org string, pattern string) *HorizonDevice {
 		Pattern: &pattern,
 		Name:    &myName,
 		Token:   &myToken,
-		HA:      &myHA,
+		HAGroup: &myHAGroup,
 	}
 	return hd
 }
