+++ exchange/pattern.go
@@ -390,59 +390,3 @@ func ConvertCommon(p *Pattern, patternId string, dv DataVerification, nodeh Node
 	}
 
 }
-
-// Structs and types for working with pattern based exchange searches
-type SearchExchangePatternRequest struct {
-	ServiceURL   string   `json:"serviceUrl,omitempty"`
-	Arch         string   `json:"arch,omitempty"`
-	NodeOrgIds   []string `json:"nodeOrgids,omitempty"`
-	SecondsStale int      `json:"secondsStale"`
-	NumEntries   int      `json:"numEntries"`
-}
-
-func (a SearchExchangePatternRequest) String() string {
-	return fmt.Sprintf("ServiceURL: %v, SecondsStale: %v, NumEntries: %v", a.ServiceURL, a.SecondsStale, a.NumEntries)
-}
-
-type SearchExchangePatternResponse struct {
-	Devices   []SearchResultDevice `json:"nodes"`
-	LastIndex int                  `json:"lastIndex"`
-}
-
-func (r SearchExchangePatternResponse) String() string {
-	return fmt.Sprintf("Devices: %v, LastIndex: %v", r.Devices, r.LastIndex)
-}
-
-// This function creates the exchange search message body.
-func CreateSearchPatternRequest() *SearchExchangePatternRequest {
-
-	ser := &SearchExchangePatternRequest{
-		NumEntries: 0,
-	}
-
-	return ser
-}
-
-func GetPatternNodes(ec ExchangeContext, policyOrg string, patternId string, req *SearchExchangePatternRequest) (*[]SearchResultDevice, error) {
-	// Invoke the exchange
-	var resp interface{}
-	resp = new(SearchExchangePatternResponse)
-	targetURL := ec.GetExchangeURL() + "orgs/" + policyOrg + "/patterns/" + GetId(patternId) + "/search"
-	for {
-		if err, tpErr := InvokeExchange(ec.GetHTTPFactory().NewHTTPClient(nil), "POST", targetURL, ec.GetExchangeId(), ec.GetExchangeToken(), *req, &resp); err != nil {
-			if !strings.Contains(err.Error(), "status: 404") {
-				return nil, err
-			} else {
-				empty := make([]SearchResultDevice, 0, 0)
-				return &empty, nil
-			}
-		} else if tpErr != nil {
-			glog.Warningf(tpErr.Error())
-			time.Sleep(10 * time.Second)
-			continue
-		} else {
-			dev := resp.(*SearchExchangePatternResponse).Devices
-			return &dev, nil
-		}
-	}
-}
