+++ exchangesync/node_policy_test.go
@@ -34,7 +34,7 @@ func Test_UpdateNodePolicy(t *testing.T) {
 	}
 	defer cleanTestDir(dir)
 
-	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "device", false, "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "device", "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("failed to create persisted device, error %v", err)
 	}
@@ -74,7 +74,7 @@ func Test_DeleteNodePolicy(t *testing.T) {
 	}
 	defer cleanTestDir(dir)
 
-	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "cluster", false, "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "cluster", "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("failed to create persisted device, error %v", err)
 	}
@@ -122,7 +122,7 @@ func Test_ExchangeNodePolicyChanged(t *testing.T) {
 	}
 	defer cleanTestDir(dir)
 
-	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "", false, "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "", "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("failed to create persisted device, error %v", err)
 	}
@@ -184,7 +184,7 @@ func Test_SetDefaultNodePolicy(t *testing.T) {
 	}
 	defer cleanTestDir(dir)
 
-	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "device", false, "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "device", "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("failed to create persisted device, error %v", err)
 	}
@@ -244,7 +244,7 @@ func Test_NodePolicyInitalSetup(t *testing.T) {
 		t.Errorf("Wrong error, should say 'Exchange registration not recorded' but got: %v", err)
 	}
 
-	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "", false, "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	pDevice, err := persistence.SaveNewExchangeDevice(db, "testid", "testtoken", "testname", "", "myOrg", "", persistence.CONFIGSTATE_CONFIGURING, persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	if err != nil {
 		t.Errorf("Unexpected error: %v", err)
 	}
