+++ microservice/microservice_test.go
@@ -197,7 +197,7 @@ func TestUnregisterServiceExchange(t *testing.T) {
 	version := "1.0"
 
 	// save device in db
-	_, err = persistence.SaveNewExchangeDevice(db, "mydevice", device_token, device_name, "device", false, org, "netspeed-amd64", "configuring", persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
+	_, err = persistence.SaveNewExchangeDevice(db, "mydevice", device_token, device_name, "device", org, "netspeed-amd64", "configuring", persistence.SoftwareVersion{persistence.AGENT_VERSION: "1.0.0"})
 	assert.Nil(t, err, fmt.Sprintf("should not return error, but got this: %v", err))
 
 	err = UnregisterMicroserviceExchange(getVariableDeviceHandler(nil, nil),
